Make this match quorum elsewhere.

+++ pkg/paxos/learner.go
@@ -6,9 +6,9 @@ const (
     lNumParts
 )
 
-func learner(quorum int, ins chan Msg) string {
+func learner(quorum uint64, ins chan Msg) string {
     var round uint64 = 0
-    votes := make(map[string]int) // maps values to number of votes
+    votes := make(map[string]uint64) // maps values to number of votes
     voted := make(map[uint64]bool) // maps values to number of votes
 
     update := func(in Msg) string {
@@ -29,7 +29,7 @@ func learner(quorum int, ins chan Msg) string {
             return ""
         case mRound > round:
             round = mRound
-            votes = make(map[string]int)
+            votes = make(map[string]uint64)
             voted = make(map[uint64]bool)
             fallthrough
         case mRound == round:
